배달 시 가게에 돈 지급 구현 및 자료형 맞춤

+++ src/main/java/com/sparta/team14project/dto/StoreResponseDto.java
@@ -12,7 +12,7 @@ public class StoreResponseDto {
     private String storeDetails;
     private String storeAddress;
     private Float avgStar;
-    private Long storePoint;
+    private int storePoint;
 
     public StoreResponseDto(Store store){
         this.id = store.getId();
